meson: add a option where to find the clang library and header
with having them people can just specify where the clang stuff is
installed, if its not possible to find it in the default location.

This fixes a bug where clang was not found, dirs: is meant as exclusive
parameter, so if you specify there a custom directory, then only THIS
specific directory will be used for the search. This fixes also that bug
and only specifies that argument if a custom location is given.