searchpanel: complete rewrite to have low hoverhead when processing search.
This search code is more efficient and will research keyword by chunk and
backward count the number of line. It will reduce the amount of random access
on disk and be access all data sequencially, page after page.

A possible next step in optimization would be to allow for the search
of multiple keyword at the same time. Another most likely bigger reward
would be to have a cache mecanism leveraging the versionning system
and only reprocessing file when they do change. This would lead to the
biggest improvement.