edi_process: Introduce new process API.
Introduced process querying API for all the platforms
we currently support. With this we are able to get
realiable data about system processes.

We use this API to track running process instances
launched from Edi.

Also, this allows for a much cleaner implementation
of the edi_debug code when tracking the debug process.

Essentially using the new API to track both running
instances and debugged instances.

Moreover, simplifying this by using a timer in
edi_main to keep track of these occurences.

We can now reliably track use of the 'Run' command
in the toolbar and menu. Adding a 'Terminate' to
compliment. The UI now responds properly to running
or non-running Edi launched process instances.