debugpanel: Improve this somewhat.
Improve the debugging process and rearrange the debug panel.

In relation to icons without text:

If this was an iPhone or Android application I can understand
representing every action with a button/icon only. If you are
developing software it's much easier to know what you are doing
with a little bit of text.