welcome: Examples.
The examples.git repo is getting RATHER large. However it is
demonstrating here that the UI can lock if we don't use a
separate thread here. Also we need some indication something
is happening, so let's use a progressbar.

This needs some more work but for now it's better behaviour.