scm: Add clone depth.
It's useful to either clone a full git history or one commit in
depth. One for cloning an existing (welcome screen) and one
shallow clone for our examples repo which is massive now!
This probably could become useful later on. It resolves the issue
of our examples being huge now.